move _createUsers test function to helper.ts file
