+++ test/Builder/SimpleBuilder.ts
@@ -4,8 +4,7 @@ import { LockDealNFT } from '../../typechain-types';
 import { LockDealProvider } from '../../typechain-types';
 import { TimedDealProvider } from '../../typechain-types';
 import { SimpleBuilder } from '../../typechain-types';
-import { deployed, token } from '../helper';
-import { _createUsers } from './BuilderHelper';
+import { deployed, token, _createUsers } from '../helper';
 import { time } from '@nomicfoundation/hardhat-network-helpers';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { expect } from 'chai';
