+++ test/Builder/SimpleRefundBuilder.ts
@@ -6,8 +6,7 @@ import { TimedDealProvider } from '../../typechain-types';
 import { CollateralProvider } from '../../typechain-types';
 import { RefundProvider } from '../../typechain-types';
 import { SimpleRefundBuilder } from '../../typechain-types';
-import { deployed, token, BUSD } from '.././helper';
-import { _createUsers } from './BuilderHelper';
+import { deployed, token, BUSD, _createUsers } from '.././helper';
 import { time } from '@nomicfoundation/hardhat-network-helpers';
 import { SignerWithAddress } from '@nomiclabs/hardhat-ethers/signers';
 import { expect } from 'chai';
